requirements.txt
pandas
python-dotenv
psycopg2-binary
requests
gensim
scikit-learn
numpy